refactor: user/dashboard 폴더로 이동
