+++ frontend/src/domains/components/FeedbackBoxList/FeedbackBoxList.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import FeedbackBoxList from './FeedbackBoxList';
-import UserFeedbackBox from '@/domains/dashboard/components/UserFeedbackBox/UserFeedbackBox';
+import UserFeedbackBox from '@/domains/user/dashboard/components/UserFeedbackBox/UserFeedbackBox';
 import AdminFeedbackBox from '@/domains/admin/home/components/AdminFeedbackBox';
 
 const meta: Meta<typeof FeedbackBoxList> = {
