+++ frontend/src/domains/user/dashboard/components/CheerButton/CheerButton.tsx
