+++ frontend/src/domains/user/dashboard/components/DashboardPanel/DashboardPanel.tsx
