+++ frontend/src/router.tsx
@@ -1,4 +1,4 @@
-import Dashboard from '@/domains/dashboard/Dashboard';
+import Dashboard from '@/domains/user/dashboard/Dashboard';
 import { createBrowserRouter } from 'react-router-dom';
 import AdminHome from './domains/admin/home/AdminHome';
 import Suggestions from './domains/user/suggestions/Suggestions';
